pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Sara Hashemi",email = "sara.hashemi@crystalloids.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk (>=1.4.2,<2.0.0)",
    "google-cloud-bigquery (>=3.34.0,<4.0.0)",
    "kagglehub (>=0.3.12,<0.4.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "pyarrow (>=20.0.0,<21.0.0)",
    "pandas-gbq (>=0.29.1,<0.30.0)",
    "immutabledict (>=4.2.1,<5.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "db-dtypes (>=1.4.3,<2.0.0)",
    "regex (>=2024.11.6,<2025.0.0)",
    "sqlglot (>=26.30.0,<27.0.0)",
    "absl-py (>=2.3.0,<3.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "jira (>=3.10.5,<4.0.0)",
    "google-cloud-discoveryengine (>=0.13.11,<0.14.0)",
    "google-cloud-aiplatform[agent-engines] (>=1.95.1,<2.0.0)",
    "flask (>=3.1.2,<4.0.0)",
    "google-auth-oauthlib (>=1.2.3,<2.0.0)",
    "google-api-python-client (>=2.187.0,<3.0.0)",
    "schedule (>=1.2.2,<2.0.0)"

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
